requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
gunicorn==21.2.0
tensorflow==2.13.0
scikit-learn==1.3.0
Pillow==10.0.1
pymongo==4.5.0
numpy==1.24.3